refactor: use Mdbx/StaticFile subcommands for db checksum
Follows the pattern used by `db get` command - adds `mdbx` and `static-file`
subcommands instead of a separate top-level command.

Usage: reth db checksum mdbx <TABLE>
  reth db checksum static-file <SEGMENT>
Amp-Thread-ID: https://ampcode.com/threads/T-019bdb7f-d203-738d-af03-55f7c5e10641
