+++ crates/cli/commands/src/db/checksum.rs
@@ -4,15 +4,17 @@ use crate::{
 };
 use alloy_primitives::map::foldhash::fast::FixedState;
 use clap::Parser;
+use itertools::Itertools;
 use reth_chainspec::EthereumHardforks;
-use reth_db::DatabaseEnv;
+use reth_db::{static_file::iter_static_files, DatabaseEnv};
 use reth_db_api::{
     cursor::DbCursorRO, table::Table, transaction::DbTx, RawKey, RawTable, RawValue, TableViewer,
     Tables,
 };
 use reth_db_common::DbTool;
 use reth_node_builder::{NodeTypesWithDB, NodeTypesWithDBAdapter};
-use reth_provider::{providers::ProviderNodeTypes, DBProvider};
+use reth_provider::{providers::ProviderNodeTypes, DBProvider, StaticFileProviderFactory};
+use reth_static_file_types::StaticFileSegment;
 use std::{
     hash::{BuildHasher, Hasher},
     sync::Arc,
@@ -23,21 +25,48 @@ use tracing::{info, warn};
 #[derive(Parser, Debug)]
 /// The arguments for the `reth db checksum` command
 pub struct Command {
-    /// The table name
-    table: Tables,
+    #[command(subcommand)]
+    subcommand: Subcommand,
+}
 
-    /// The start of the range to checksum.
-    #[arg(long, value_parser = maybe_json_value_parser)]
-    start_key: Option<String>,
+#[derive(clap::Subcommand, Debug)]
+enum Subcommand {
+    /// Calculates the checksum of a database table
+    Mdbx {
+        /// The table name
+        table: Tables,
 
-    /// The end of the range to checksum.
-    #[arg(long, value_parser = maybe_json_value_parser)]
-    end_key: Option<String>,
+        /// The start of the range to checksum.
+        #[arg(long, value_parser = maybe_json_value_parser)]
+        start_key: Option<String>,
 
-    /// The maximum number of records that are queried and used to compute the
-    /// checksum.
-    #[arg(long)]
-    limit: Option<usize>,
+        /// The end of the range to checksum.
+        #[arg(long, value_parser = maybe_json_value_parser)]
+        end_key: Option<String>,
+
+        /// The maximum number of records that are queried and used to compute the
+        /// checksum.
+        #[arg(long)]
+        limit: Option<usize>,
+    },
+    /// Calculates the checksum of a static file segment
+    StaticFile {
+        /// The static file segment
+        #[arg(value_enum)]
+        segment: StaticFileSegment,
+
+        /// The block number to start from (inclusive).
+        #[arg(long)]
+        start_block: Option<u64>,
+
+        /// The block number to end at (inclusive).
+        #[arg(long)]
+        end_block: Option<u64>,
+
+        /// The maximum number of rows to checksum.
+        #[arg(long)]
+        limit: Option<usize>,
+    },
 }
 
 impl Command {
@@ -47,12 +76,107 @@ impl Command {
         tool: &DbTool<NodeTypesWithDBAdapter<N, Arc<DatabaseEnv>>>,
     ) -> eyre::Result<()> {
         warn!("This command should be run without the node running!");
-        self.table.view(&ChecksumViewer {
-            tool,
-            start_key: self.start_key,
-            end_key: self.end_key,
-            limit: self.limit,
-        })?;
+
+        match self.subcommand {
+            Subcommand::Mdbx { table, start_key, end_key, limit } => {
+                table.view(&ChecksumViewer { tool, start_key, end_key, limit })?;
+            }
+            Subcommand::StaticFile { segment, start_block, end_block, limit } => {
+                self.checksum_static_file(tool, segment, start_block, end_block, limit)?;
+            }
+        }
+
+        Ok(())
+    }
+
+    fn checksum_static_file<N: CliNodeTypes<ChainSpec: EthereumHardforks>>(
+        &self,
+        tool: &DbTool<NodeTypesWithDBAdapter<N, Arc<DatabaseEnv>>>,
+        segment: StaticFileSegment,
+        start_block: Option<u64>,
+        end_block: Option<u64>,
+        limit: Option<usize>,
+    ) -> eyre::Result<()> {
+        let static_file_provider = tool.provider_factory.static_file_provider();
+        if let Err(err) =
+            static_file_provider.check_consistency(&tool.provider_factory.provider()?)
+        {
+            warn!("Error checking consistency of static files: {err}");
+        }
+
+        let static_files = iter_static_files(static_file_provider.directory())?;
+
+        let ranges = static_files
+            .get(&segment)
+            .ok_or_else(|| eyre::eyre!("No static files found for segment: {}", segment))?;
+
+        let start_time = Instant::now();
+        let mut hasher = FixedState::with_seed(u64::from_be_bytes(*b"RETHRETH")).build_hasher();
+        let mut total_rows = 0usize;
+        let limit = limit.unwrap_or(usize::MAX);
+
+        let start_block = start_block.unwrap_or(0);
+        let end_block = end_block.unwrap_or(u64::MAX);
+
+        info!(
+            "Computing checksum for {} static files, start_block={}, end_block={}, limit={:?}",
+            segment,
+            start_block,
+            end_block,
+            if limit == usize::MAX { None } else { Some(limit) }
+        );
+
+        for (block_range, _header) in ranges.iter().sorted_by_key(|(range, _)| range.start()) {
+            if block_range.end() < start_block || block_range.start() > end_block {
+                continue;
+            }
+
+            let fixed_block_range =
+                static_file_provider.find_fixed_range(segment, block_range.start());
+            let jar_provider = static_file_provider
+                .get_segment_provider_for_range(segment, || Some(fixed_block_range), None)?
+                .ok_or_else(|| {
+                    eyre::eyre!(
+                        "Failed to get segment provider for segment {} at range {}",
+                        segment,
+                        block_range
+                    )
+                })?;
+
+            let mut cursor = jar_provider.cursor()?;
+
+            while let Ok(Some(row)) = cursor.next_row() {
+                for col_data in row.iter() {
+                    hasher.write(col_data);
+                }
+
+                total_rows += 1;
+
+                if total_rows.is_multiple_of(100_000) {
+                    info!("Hashed {total_rows} rows.");
+                }
+
+                if total_rows >= limit {
+                    break;
+                }
+            }
+
+            drop(jar_provider);
+            static_file_provider.remove_cached_provider(segment, fixed_block_range.end());
+
+            if total_rows >= limit {
+                break;
+            }
+        }
+
+        let checksum = hasher.finish();
+        let elapsed = start_time.elapsed();
+
+        info!(
+            "Checksum for static file segment `{}`: {:#x} ({} rows, elapsed: {:?})",
+            segment, checksum, total_rows, elapsed
+        );
+
         Ok(())
     }
 }
